scripts/requirements_dashboard.txt
# Requirements for dashboard script
google-cloud-bigquery==3.15.0
pandas==2.1.4
plotly==5.18.0

